grobid/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = 'l3hiai-pdf2text'
version = '0.1.0'
license = {text = 'Proprietary License'}
description = 'GROBID-backed PDF2Text Utility'
authors = [
    {name="Fitzroy Nembhard", email="fnembhard@fit.edu"},
    {name="Marco Carvalho", email="mcarvalho@fit.edu"}
]
maintainers = [
    {name="Fitzroy Nembhard", email="fnembhard@fit.edu"},
]

readme = "README.md"

classifiers=["Environment :: Console"

]

dependencies=[
       'doi2pdf==0.1.3',
        'grobid_client_python==0.0.8',
        'ipython==8.12.3',
        'langchain==0.2.16',
        'langchain_community==0.2.16',
        'lxml==4.9.1',
        'pandas==2.2.2',
        'Requests==2.32.3',
        'setuptools==66.0.0',
        'SQLAlchemy==2.0.31'
    ]

[project.scripts]
l3hiai-pdf2text-cli = "l3hiai_pdf2text:main"

[tool.setuptools.packages.find]
where = ["src"]